requirements.txt
aiodns==3.1.1
aiofiles==23.1.0
aiogram==2.25.2
aiogram-utils==0.1.4
aiohttp==3.8.6
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
astroid==3.0.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.6.2
briefcase==0.3.16
Brotli==1.1.0
browser-cookie3==0.19.1
build==1.0.3
buildozer==1.5.0
cachetools==5.3.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.5.0
cssselect2==0.7.0
curl_cffi==0.6.0b9
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.7
dnspython==2.5.0
docstring-inheritance==2.1.2
duckduckgo_search==4.4
fastapi==0.109.2
filelock==3.12.4
Flask==2.3.3
flask-babel==4.0.0
frozenlist==1.4.0
g4f==0.2.6.2
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gspread==5.12.0
h11==0.14.0
idna==3.4
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
loguru==0.7.2
lxml==4.9.3
lz4==4.3.3
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==4.1.0
mongoengine==0.24.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.25.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
pep8==1.7.1
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.6
ptyprocess==0.7.0
py-arkose-generator==0.0.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycares==4.4.0
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.3.0
pydantic_core==2.6.3
PyExecJS==1.5.1
pygame==2.5.2
Pygments==2.17.2
pylint==3.0.3
pymongo==4.6.1
pyproject_hooks==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
pyTelegramBotAPI==4.15.4
python-dateutil==2.8.2
python-slugify==8.0.1
python-socks==2.4.4
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
selenium==4.17.2
sh==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
spotipy==2.23.0
starlette==0.36.3
sympy==1.12
telebot==0.0.5
telepot==12.7
text-unidecode==1.3
tinycss2==1.2.1
tomli_w==1.0.0
tomlkit==0.12.3
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
undetected-chromedriver==3.5.4
urllib3==2.0.7
uvicorn==0.27.0.post1
vboxapi==1.0
virtualenv==20.24.6
webencodings==0.5.1
websockets==12.0
Werkzeug==2.3.7
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.9.2
